requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
spotipy==2.25.1
urllib3==2.5.0
